tests: Avoid null passed to unlink
When tests are skipped, there is a deprecation notice

Deprecated: unlink(): Passing null to parameter #1 ($filename) of type
string is deprecated in tests\CdbTest.php on line 34

+++ tests/CdbTest.php
@@ -31,8 +31,12 @@ class CdbTest extends TestCase {
 	}
 
 	protected function tearDown(): void {
-		unlink( $this->phpCdbFile );
-		unlink( $this->dbaCdbFile );
+		if ( $this->phpCdbFile !== null ) {
+			unlink( $this->phpCdbFile );
+		}
+		if ( $this->dbaCdbFile !== null ) {
+			unlink( $this->dbaCdbFile );
+		}
 		parent::tearDown();
 	}
